feat: add maximize button for view-mode variables
