+++ operate/client/src/App/ProcessInstance/BottomPanelTabs/VariablesTab/Variables/ViewFullVariableButton/types.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+export interface ViewFullVariableButtonProps {
+  variableName: string;
+  variableValue: string;
+  variableKey: string;
+}
+
+export interface ViewFullVariableWrapperProps extends ViewFullVariableButtonProps {
+  isEditMode: boolean;
+}
